Fix playback offset for certain MP3 files
Fixes #187

Here are my notes that I wrote down while debugging this:

file: Loop.mp3

startSilenceSamples: 1105

The same reference sample:
- Audacity: 1734
- NAudio: 2263

Difference: 529 samples

When we view waveform in Tempora, we see this sample appear at 0:051, which is exactly equal to 2263/sampleRate

In other words, the timestamp we see in Tempora is exactly equal to how NAudio understands the audio, sample by sample.

When we seek playback at 0:051, the current implementation does its best to make Godot's audio player play from the correct time,
by subtracting the supposed playback origin from the sampleTime. In other words, it starts playback at what Godot understands as sample 1158, or 0.0262 seconds.

However, in this file, the actual sample is timed correctly when it is lined up with the waveform's 0.077
This is around sample 3396

This means that with the current configuration, you tell Godot to play at 0.077-0.025 = 0.051 - which is what Godot's playback understands as where this sample is.
In other words, the playbackOrigin_Seconds might actually be 0 in this case???

I will try to re-launch the audio file and see what I get when I initialize the audio (by looking at if/how the Xing header is read)

It turns out that this audio file does not have a Xing header. I will assume that this means to always use a playback offset of 0 when that's the case.